src/image_collector/requirements.txt
requests
python-dotenv
geojson
shapely
rich
pandas
geopandas
APScheduler>=3.11.0
PyYAML>=6.0
python-daemon>=3.0